wip: Refactor summary data into summary state
